+++ packages/repo-app-import-sub-generator/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @sap-ux/repo-app-download-sub-generator
 
+## 0.3.76
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.5.20
+-   @sap-ux/odata-service-inquirer@2.5.29
+
+## 0.3.75
+
+### Patch Changes
+
+-   Updated dependencies [9f10a60]
+    -   @sap-ux/odata-service-inquirer@2.5.29
+    -   @sap-ux/axios-extension@1.22.5
+    -   @sap-ux/fiori-generator-shared@0.13.6
+    -   @sap-ux/system-access@0.6.12
+    -   @sap-ux/abap-deploy-config-writer@0.2.5
+
+## 0.3.74
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-writer@0.2.4
+-   @sap-ux/fiori-elements-writer@2.5.19
+-   @sap-ux/launch-config@0.10.17
+-   @sap-ux/project-access@1.30.10
+-   @sap-ux/odata-service-inquirer@2.5.28
+-   @sap-ux/axios-extension@1.22.4
+-   @sap-ux/fiori-generator-shared@0.13.6
+-   @sap-ux/project-input-validator@0.6.15
+-   @sap-ux/system-access@0.6.11
+-   @sap-ux/inquirer-common@0.7.31
+
+## 0.3.73
+
+### Patch Changes
+
+-   Updated dependencies [9fa7f0b]
+    -   @sap-ux/odata-service-inquirer@2.5.27
+    -   @sap-ux/inquirer-common@0.7.30
+
 ## 0.3.72
 
 ### Patch Changes