+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.14.35
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.10
+-   @sap-ux/project-input-validator@0.6.15
+-   @sap-ux/inquirer-common@0.7.31
+
+## 0.14.34
+
+### Patch Changes
+
+-   Updated dependencies [9fa7f0b]
+    -   @sap-ux/inquirer-common@0.7.30
+
 ## 0.14.33
 
 ### Patch Changes