+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.1.35
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-writer@0.6.12
+-   @sap-ux/fiori-generator-shared@0.13.6
+-   @sap-ux/ui5-library-inquirer@0.4.34
+
+## 0.1.34
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-inquirer@0.4.33
+
 ## 0.1.33
 
 ### Patch Changes